remove all dep code for 3.1
