+++ changelog/5304.removal.1.rst
@@ -1,0 +1,1 @@
+``sunpy.roi.chaincode.Chaincode`` has been removed in favour of `sunpy.net.helio.Chaincode`.
