+++ changelog/5304.removal.2.rst
@@ -1,0 +1,1 @@
+Removed ``sunpy.roi.roi``, there is no direct replacement but `astropy-regions <https://astropy-regions.readthedocs.io/en/latest/>`__ is something to consider.
