+++ changelog/5304.removal.5.rst
@@ -1,0 +1,1 @@
+Removed the ability to read txt files from `NOAAIndicesTimeSeries` and `NOAAPredictIndicesTimeSeries` as the data provided by NOAA is now provided as JSON files.
