+++ changelog/5304.removal.7.rst
@@ -1,0 +1,1 @@
+Removed ``sunpy.map.GenericMap.size``, please use ``sunpy.map.GenericMap.data.size``.
