--- docs/code_ref/instr.rst
@@ -1,22 +1,0 @@
-SunPy instr
-***********
-
-The ``instr`` submodule contains instrument specific routines, generally this
-submodule is focused on routines that process calibrated data.
-
-.. warning::
-    This module will be removed in sunpy 3.1.
-    This code has been moved out to `sunkit-instruments <https://pypi.org/project/sunkit-instruments/>`__.
-
-
-.. automodapi:: sunpy.instr
-
-.. automodapi:: sunpy.instr.fermi
-
-.. automodapi:: sunpy.instr.goes
-
-.. automodapi:: sunpy.instr.iris
-
-.. automodapi:: sunpy.instr.lyra
-
-.. automodapi:: sunpy.instr.rhessi
