+++ setup.cfg
@@ -46,10 +46,6 @@ dask =
   dask[array]>=2.0.0
 database =
   sqlalchemy>=1.3.4
-instr =
-  matplotlib>=3.1.0
-  pandas>=0.24.0
-  scipy>=1.3.0
 image =
   scikit-image>=0.16.0
   scipy>=1.3.0
@@ -178,8 +174,6 @@ filterwarnings =
     # Recent change to astropy has made our figure tests emit this warning.
     # https://github.com/sunpy/sunpy/issues/5073
     ignore:'datfix' made the change:astropy.wcs.wcs.FITSFixedWarning
-    # Ignore internal instr dep warning until we remove it
-    ignore:sunpy.instr is deprecated and will be removed in sunpy 3.1.
 
 [pycodestyle]
 max_line_length = 110
