--- sunpy/extern/sunkit_instruments/fermi/__init__.py
@@ -1,1 +1,0 @@
-from .fermi import *  # NOQA
