--- sunpy/extern/sunkit_instruments/goes_xrs/__init__.py
@@ -1,1 +1,0 @@
-from .goes_xrs import *  # NOQA
