--- sunpy/extern/sunkit_instruments/iris/__init__.py
@@ -1,1 +1,0 @@
-from .iris import *  # NOQA
