--- sunpy/extern/sunkit_instruments/lyra/__init__.py
@@ -1,1 +1,0 @@
-from .lyra import *  # NOQA
