--- sunpy/extern/sunkit_instruments/rhessi/__init__.py
@@ -1,1 +1,0 @@
-from .rhessi import *  # NOQA
