--- sunpy/instr/__init__.py
@@ -1,7 +1,0 @@
-import warnings
-
-from sunpy.util.exceptions import SunpyDeprecationWarning
-
-warnings.warn("sunpy.instr is deprecated and will be removed in sunpy 3.1. "
-              "Please use the affiliated sunkit-instruments package.",
-              SunpyDeprecationWarning)
