--- sunpy/instr/rhessi.py
@@ -1,15 +1,0 @@
-from sunpy.extern.sunkit_instruments.rhessi import (
-    backprojection,
-    parse_observing_summary_dbase_file,
-    parse_observing_summary_hdulist,
-)
-
-__all__ = ['parse_observing_summary_hdulist',
-           'backprojection',
-           'parse_observing_summary_dbase_file']
-
-# Trick the docs into thinking these functions are defined in here.
-for _a in (backprojection,
-           parse_observing_summary_dbase_file,
-           parse_observing_summary_hdulist):
-    _a.__module__ = __name__
