--- sunpy/roi/__init__.py
@@ -1,2 +1,0 @@
-from sunpy.roi.chaincode import *
-from sunpy.roi.roi import *
