+++ sunpy/tests/figure_hashes_mpl_332_ft_261_astropy_42.json
@@ -46,9 +46,7 @@
   "sunpy.timeseries.tests.test_timeseriesbase.test_lyra_peek": "d134646089aa9cb217b798e61fbfbfdcf6acf50a18f9292e95294d2cae6d24ea",
   "sunpy.timeseries.tests.test_timeseriesbase.test_rhessi_peek": "ffd453337e44d613db520d0bf64966d812223f4e3bec8222d6fb7e4cbc4b4756",
   "sunpy.timeseries.tests.test_timeseriesbase.test_noaa_json_ind_peek": "1bea3a56178db101abca168bc51e127b7b96f9eed58d74e3d19b6ffc14643e80",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_noaa_txt_ind_peek": "01fa7e5baa13f996f686d18505f177f0884deb394384a129522aa18f9c6b61ac",
   "sunpy.timeseries.tests.test_timeseriesbase.test_noaa_json_pre_peek": "4c4a4a299ed04547a70ea8da89871e05a897d5f8937d5cefc0359d04cf079c04",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_noaa_txt_pre_peek": "fbda770f164bd38a7741d3393e9f93040b188e928521cc53379ba2b78beb582c",
   "sunpy.timeseries.tests.test_timeseriesbase.test_generic_ts_peek": "73a71e717b7dd0dd345f0f8a8bfa7f06387607417b3ecd123adfd8f630f93d51",
   "sunpy.visualization.animator.tests.test_basefuncanimator.test_lineanimator_figure": "dbf47ab983a844621f912b2f5a2d9ee657d3f661678c16a07afc7217f50e3fd4",
   "sunpy.visualization.animator.tests.test_mapsequenceanimator.test_map_sequence_animator_wcs_simple_plot": "8399e8f0c46655b2d5e868cbf81603abf9f2f56157d94f2be4056ee9ababb237",
